+++ src/main/java/com/ametrinstudios/ametrin/util/TimeHelper.java
@@ -1,6 +1,9 @@
 package com.ametrinstudios.ametrin.util;
 
+@SuppressWarnings("unused")
+
 public final class TimeHelper {
     public static int SecondsToTicks(int seconds) { return seconds * 20; }
     public static int MinutesToTicks(int minutes) { return SecondsToTicks(minutes * 60); }
+    public static int HoursToTicks(int hours) { return MinutesToTicks(hours * 60); }
 }